[Update]ダメージ関連の処理をintで処理をするように変更

+++ Assets/Scripts/Charactor/CharactorStatus.cs
@@ -10,5 +10,5 @@ public class CharacterStatus : ScriptableObject
     public string Name;
     public int MaxHp;
     public float MoveSpeed;
-    public float AttackPower;
+    public int AttackPower;
 }
